Made type value a long.

+++ Prog0/src/sdns/serialization/A.java
@@ -85,7 +85,7 @@ public class A extends ResourceRecord {
      * @return type value
      */
     @Override
-    public long getTypeValue() { return A_TYPE_VALUE; }
+    public int getTypeValue() { return A_TYPE_VALUE; }
 
     /**
      * Returns a byte array of the rdata for this object.  For internal use only.
